on/off

+++ front/views/symbols.php
@@ -67,7 +67,6 @@
 		width: 10%;
 	}
 
-	.save-btn,
 	.delete-btn {
 		width: 15vh;
 		margin: 0.2vh;
@@ -81,7 +80,6 @@
 		border-radius: 1vh;
 	}
 
-	.save-btn:hover,
 	.delete-btn:hover {
 		border: 2px inset grey;
 	}
@@ -106,6 +104,44 @@
 		outline: none;
 		border: 1px solid grey;
 		border-radius: 0.2vh;
+	}
+
+
+	/* Style pour le bouton ON/OFF avec curseur glissant type Apple */
+	.toggle-btn {
+		position: relative;
+		display: inline-block;
+		width: 4em;
+		height: 2em;
+		background-color: #ddd;
+		border-radius: 1em;
+		cursor: pointer;
+		transition: background-color 0.3s ease;
+		overflow: hidden;
+	}
+
+	.toggle-btn::before {
+		content: '';
+		position: absolute;
+		top: 50%;
+		transform: translateY(-50%);
+		left: 0.3em;
+		width: 1.4em;
+		height: 1.4em;
+		background-color: rgba(255, 255, 255, 0.5);
+		border-radius: 50%;
+		border: 1px solid grey;
+		box-shadow: 0 0.1em 0.2em rgba(0, 0, 0, 0.2);
+		transition: transform 0.3s ease;
+	}
+
+	.toggle-btn.active {
+		background-color: #4cd964;
+		color: white;
+	}
+
+	.toggle-btn.active::before {
+		transform: translateX(1.8em) translateY(-50%);
 	}
 
 	/* Mise en page pour les catégories */
@@ -257,7 +293,7 @@
 		<label for="categoryFilter">Filtrer par catégorie :</label>
 		<select id="categoryFilter">
 			<option value="">Toutes les catégories</option>
-			<!-- Les options de catégorie seront ajoutées ici par JavaScript -->
+			<!-- Les options de catégorie seront ajoutées ici en JavaScript -->
 		</select>
 	</div>
 
@@ -272,7 +308,7 @@
 		<th>Mots-clés</th>
 		<th>Actions</th>
 	</tr>
-	<!-- Les lignes du tableau seront ajoutées ici par JavaScript -->
+	<!-- Les lignes du tableau seront ajoutées ici en JavaScript -->
 </table>
 
 
@@ -332,10 +368,10 @@
 			</div>
 		  </td>
 
-		  <td class="cln-size contenteditable" data-field="active">
-			<div class="editable-content" contenteditable="true">
-			  ${symbol.active}
-			</div>
+		  <td class="cln-size" data-field="active">
+			<button class="toggle-btn ${symbol.active ? 'active' : ''}" data-id="${symbol.id}">
+			  ${symbol.active ? 'ON' : 'OFF'}
+			</button>
 		  </td>
 
 		  <td>
@@ -369,7 +405,6 @@
 		  </td>
 
 		  <td>
-			<button class="save-btn" data-id="${symbol.id}">Enregistrer</button>
 			<button class="delete-btn" data-id="${symbol.id}">Supprimer</button>
 		  </td>
 		</tr>
@@ -378,6 +413,7 @@
 
 		$('.table-symbols').append(tableContent);
 	});
+
 
 	// _____________________________MODAL__________________________________
 
@@ -472,7 +508,7 @@
 
 	function addItem(itemElement, itemType) {
 		var symbolRow = itemElement.closest("tr");
-		var symbolId = symbolRow.querySelector(".save-btn").dataset.id;
+		var symbolId = symbolRow.querySelector(".delete-btn").dataset.id;
 		var itemId = itemElement.dataset[itemType + "Id"];
 		var apiUrl = apiBaseURL + "/symbols/" + symbolId;
 
@@ -522,12 +558,12 @@
 		});
 	}
 
-	// ___________________________________________________________________________
+	// __________________________________ Supprimer la liaison categories ou mots-clés sur les symboles _________________________________________
 	function deleteCategory() {
 		var categoryItem = this.closest("li");
 		var categoryId = categoryItem.dataset.categoryId;
 		var symbolRow = this.closest("tr");
-		var symbolId = symbolRow.querySelector(".save-btn").dataset.id;
+		var symbolId = symbolRow.querySelector(".delete-btn").dataset.id;
 
 		// Effectuer une requête AJAX DELETE pour supprimer la catégorie du symbole
 		$.ajax({
@@ -547,7 +583,7 @@
 		var keywordItem = this.closest("li");
 		var keywordId = keywordItem.dataset.keywordId;
 		var symbolRow = this.closest("tr");
-		var symbolId = symbolRow.querySelector(".save-btn").dataset.id;
+		var symbolId = symbolRow.querySelector(".delete-btn").dataset.id;
 
 		// Effectuer une requête AJAX DELETE pour supprimer le mot-clé du symbole
 		$.ajax({
@@ -563,26 +599,57 @@
 		});
 	}
 
-	// Attacher des gestionnaires d'événements aux éléments du tableau
-	$(document).on("click", ".add-category-btn", openModal);
-	$(document).on("click", ".add-keyword-btn", openModal);
-	$(document).on("click", ".delete-category-btn", deleteCategory);
-	$(document).on("click", ".delete-keyword-btn", deleteKeyword);
+	//________________ Fonction pour gérer le basculement du bouton ON/OFF ______________
+	function toggleActive(symbolId) {
+		var toggleButton = $(`.toggle-btn[data-id="${symbolId}"]`);
+		var isActive = toggleButton.hasClass("active") ? 0 : 1; // Convertir la valeur en entier
 
-	// Attacher des gestionnaires d'événements pour enregistrer modifications et supprimer les symboles
-	$(document).on("click", ".save-btn", function () {
+		// Effectuer une requête AJAX PUT pour mettre à jour l'état actif du symbole
+		$.ajax({
+			url: apiBaseURL + "/symbols/" + symbolId,
+			type: "PUT",
+			data: JSON.stringify({
+				active: isActive === 1 ? 1 : 0 // Assurer que la valeur est soit 1 soit 0
+			}),
+			contentType: "application/json",
+			success: function (data) {
+				console.log("État actif du symbole mis à jour avec succès:", data);
+				// Mettre à jour le bouton ON/OFF et son état actif
+				toggleButton.toggleClass("active");
+				toggleButton.text(toggleButton.hasClass("active") ? "ON" : "OFF");
+
+				// Mettre à jour le symbole dans le tableau
+				var symbolRow = toggleButton.closest("tr");
+				updateSymbol(symbolId, symbolRow);
+			},
+			error: function (jqXHR, textStatus, errorThrown) {
+				console.log("Erreur lors de la mise à jour de l'état actif du symbole:", textStatus, errorThrown);
+			}
+		});
+	}
+
+	// Gestionnaire d'événement pour le basculement de l'état actif
+	$(document).on("click", ".toggle-btn", function () {
 		var symbolId = $(this).data("id");
-		var symbolRow = $(this).closest("tr");
-		var fileName = symbolRow.find("[data-field='file_name'] .editable-content").text().trim();
-		var size = symbolRow.find("[data-field='size'] .editable-content").text().trim();
-		var active = symbolRow.find("[data-field='active'] .editable-content").text().trim();
+		var isActive = $(this).hasClass("active");
 
-		console.log(symbolRow);
+		toggleActive(symbolId, !isActive);
+	});
+
+	//_____________________ UPDATE SYMBOLE ____________________________________
+	// Fonction pour gérer la mise à jour du symbole
+	function updateSymbol(symbolId, symbolRow) {
+		var fileNameElement = symbolRow.find("[data-field='file_name'] .editable-content");
+		var sizeElement = symbolRow.find("[data-field='size'] .editable-content");
+		var activeElement = symbolRow.find("[data-field='active'] .toggle-btn");
+
+		var fileName = fileNameElement.text().trim();
+		var size = parseInt(sizeElement.text().trim());
+		var active = activeElement.hasClass("active") ? 1 : 0;
 
 		var categories = Array.from(symbolRow[0].querySelectorAll(".category-list li[data-category-id]")).map(function (category) {
 			return parseInt(category.dataset.categoryId);
 		});
-		console.log(categories);
 
 		var keywords = Array.from(symbolRow[0].querySelectorAll(".keyword-list li")).map(function (keyword) {
 			return parseInt(keyword.dataset.keywordId);
@@ -594,8 +661,8 @@
 			type: "PUT",
 			data: JSON.stringify({
 				file_name: fileName,
-				size: parseInt(size),
-				active: parseInt(active),
+				size: size,
+				active: active,
 				categories: categories,
 				keywords: keywords
 			}),
@@ -607,24 +674,52 @@
 				console.log("Erreur lors de la mise à jour du symbole:", textStatus, errorThrown);
 			}
 		});
+	}
+
+	// Attacher un gestionnaire d'événement à la modification du champ "Nom du fichier"
+	$(document).on("input", "[data-field='file_name'] .editable-content", function () {
+		var symbolId = $(this).closest("tr").find(".delete-btn").data("id");
+		var symbolRow = $(this).closest("tr");
+		updateSymbol(symbolId, symbolRow);
 	});
 
+	// Attacher un gestionnaire d'événement à la modification du champ "Taille"
+	$(document).on("input", "[data-field='size'] .editable-content", function () {
+		var symbolId = $(this).closest("tr").find(".delete-btn").data("id");
+		var symbolRow = $(this).closest("tr");
+		updateSymbol(symbolId, symbolRow);
+	});
+
+
+
+	//___________________ gestionnaires d'événements aux éléments du tableau _______________________
+	$(document).on("click", ".add-category-btn", openModal);
+	$(document).on("click", ".add-keyword-btn", openModal);
+	$(document).on("click", ".delete-category-btn", deleteCategory);
+	$(document).on("click", ".delete-keyword-btn", deleteKeyword);
+
+
+
+	//__________________ SUPPRESSION D'UN SYMBOLE ______________________
 	$(document).on("click", ".delete-btn", function () {
 		var symbolId = $(this).data("id");
 		var symbolRow = $(this).closest("tr");
 
-		// Effectuer une requête AJAX DELETE pour supprimer le symbole
-		$.ajax({
-			url: apiBaseURL + "/symbols/" + symbolId,
-			type: "DELETE",
-			success: function () {
-				console.log("Symbole supprimé avec succès.");
-				symbolRow.remove();
-			},
-			error: function (jqXHR, textStatus, errorThrown) {
-				console.log("Erreur lors de la suppression du symbole:", textStatus, errorThrown);
-			}
-		});
+		// Afficher une boîte de dialogue de confirmation avant la suppression
+		if (confirm("Êtes-vous sûr de vouloir supprimer ce symbole ?")) {
+			// Effectuer une requête AJAX DELETE pour supprimer le symbole
+			$.ajax({
+				url: apiBaseURL + "/symbols/" + symbolId,
+				type: "DELETE",
+				success: function () {
+					console.log("Symbole supprimé avec succès.");
+					symbolRow.remove();
+				},
+				error: function (jqXHR, textStatus, errorThrown) {
+					console.log("Erreur lors de la suppression du symbole:", textStatus, errorThrown);
+				}
+			});
+		}
 	});
